+++ projects/ids-enterprise-ng/src/lib/datagrid/soho-datagrid.component.ts
@@ -169,6 +169,7 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
     this._gridOptions = gridOptions;
 
     this.checkForComponentEditors();
+    this.checkForSummaryRowSettings();
 
     if (this.jQueryElement) {
       // No need to set the 'settings' as the Rebuild will create
@@ -489,8 +490,7 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
 
   /**
    * Sets the row height for the grid, to be one of the supported options.
-   *
-   * @param rowHeight - 'normal' | 'medium' | 'short'
+   * @param rowHeight - 'extra-small' | 'small' | 'medium' | 'large'
    */
   @Input() set rowHeight(rowHeight: SohoDataGridRowHeight) {
     this._gridOptions.rowHeight = rowHeight;
@@ -498,7 +498,6 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
       this.datagrid.settings.rowHeight = rowHeight;
 
       this.ngZone.runOutsideAngular(() => {
-        // @todo add hints as this may be bundled up with other changes.
         this.datagrid.rowHeight(rowHeight);
       });
     }
@@ -1000,6 +999,19 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
         // @todo add hints for this too, as other changes may force a rebuild?
         this.datagrid.updateColumns(this._gridOptions.columns, this._gridOptions.columnGroups);
       });
+    }
+  }
+
+  /**
+   *
+   * Summary row columns settingss
+   */
+  @Input() set summaryRowColumns(summaryRowColumns: SohoDataGridSummaryRowColumnSettings[]) {
+    this._gridOptions.summaryRowColumns = summaryRowColumns;
+    this.checkForSummaryRowSettings();
+
+    if (this._gridOptions.columns && this.jQueryElement) {
+      this.ngZone.runOutsideAngular(() => this.datagrid.updateColumns(this._gridOptions.columns, this._gridOptions.columnGroups));
     }
   }
 
@@ -2624,6 +2636,26 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
       }
     });
   }
+
+  private checkForSummaryRowSettings() {
+    if (!this._gridOptions.summaryRowColumns || this._gridOptions.summaryRowColumns.length === 0 ) {
+      this._gridOptions.columns.forEach((c) => {
+        c.summaryRowFormatter = undefined;
+        c.summaryText = undefined;
+        c.aggregator = undefined;
+        c.summaryTextPlacement = undefined;
+      });
+    } else {
+      this._gridOptions.summaryRowColumns.forEach((sc) => {
+        const column = this._gridOptions.columns.find((c) => c.field === sc.field);
+        column.summaryRowFormatter = sc.summaryRowFormatter;
+        column.summaryText = sc.summaryText;
+        column.aggregator = sc.aggregator;
+        column.summaryTextPlacement = sc.summaryTextPlacement;
+      });
+    }
+  }
+
 }
 
 /**